leptos 3 integration works as i want it to now.
the diff thing doesn't care about saving in a smart way, it just uses leptos default batching
to avoid "spamming every input as a new diff to update to the db" but it works good enough
for now.

my code loads from the local db on load, and updates on diff and on reorder properly.

i commented out some "unused warning" stuff in wasm4 when i added so i can get_data_by_order
and that might have broken other  code that uses the thing because i think the use stuff
was for other projects? not a good dependency for those projects to use a dependency imported
by the the wasm4 project that the wasm4 project doesn't use. but anyway. just noting that down
so i technically have it somewhere?

next is copying leptos 3 integration project with yrs. so that's next
yrs_integration.

+++ text_diff2/src/text_block.rs
@@ -3,12 +3,12 @@ use leptos::prelude::*;
 #[derive(Copy, Clone)]
 pub struct TextBlock {
     pub text: (ReadSignal<String>, WriteSignal<String>),
-    pub id: RwSignal<usize>,
+    pub id: RwSignal<f64>,
     pub latest_diff: RwSignal<String>,
 }
 
 impl TextBlock {
-    pub fn new(text: String, id: RwSignal<usize>) -> Self {
+    pub fn new(text: String, id: RwSignal<f64>) -> Self {
         Self {
             text: signal(text),
             id,
